pyproject.toml
[project]
name = "pdf-translator"
version = "0.1.0"
description = "学術論文PDFのレイアウト維持翻訳ツール"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf",
    "google-genai",
    "python-dotenv",
]

[tool.uv]
required-environments = ["sys_platform == 'darwin' and platform_machine == 'arm64'"]
